Moved to opto/features/optimize_anything and updated related imports and tests
